refactored most of the download/upload/thumbnail code. removed depth limit for categories/folders. changes does not fully tested yet.

+++ py_files/generate_thumb.py
@@ -2,40 +2,102 @@ from json import load
 from subprocess import check_output, run
 from sys import argv
 from threading import Thread
-from os import path as os_path, makedirs
+from os import path as os_path, makedirs, _exit
 from cv2 import VideoCapture, imwrite, CAP_PROP_FRAME_COUNT, CAP_PROP_POS_FRAMES, IMWRITE_WEBP_QUALITY
 
+ABORT = 0
 
-FFMPEG = "ffmpeg -loglevel error"
-EXTRA_WEBP = "-quality 1 -compression_level 6 -speed 6"
-EXTRA_MP4 = "-crf 50 -b:v 50k -preset veryfast"
-CV2_WEBP = 1
+def abort():
+	global ABORT
+	ABORT = 1
+	global PRINT_CALLBACK
+	PRINT_CALLBACK("aborting..")
+	exit()
+	_exit()
+
+def start(url, weburl, ftype, print_callback=print):
+	global ABORT
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
+	global FFMPEG
+	global EXTRA_WEBP
+	global EXTRA_MP4
+	global CV2_WEBP
+	FFMPEG = "ffmpeg -loglevel error"
+	EXTRA_WEBP = "-quality 1 -compression_level 6 -speed 6"
+	EXTRA_MP4 = "-crf 50 -b:v 50k -preset veryfast"
+	CV2_WEBP = 1
+
+	global URL
+	global WEBURL
+	global FTYPE
+	global DUR
+	global OUTDIR
+	global APIKEY
+	global PRINT_CALLBACK
+	URL = fr"{url}"
+	WEBURL = weburl
+	FTYPE = ftype
+	DUR = getdur(URL)
+	randomname = str(abs(hash(URL)))
+	with open(r"./config/config.json", "r") as f:
+		j = load(f)
+		OUTDIR = fr'{j["tmp_dir"]}/{randomname}'
+		APIKEY = j["api_key"]
+	if not os_path.exists(OUTDIR):
+		makedirs(OUTDIR)
+	
+	PRINT_CALLBACK = print_callback
+
+	PRINT_CALLBACK(f"url: {URL} && got dur: {DUR} && && outdir: {OUTDIR}")
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
+	generate()
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
+	extract_middle_frame()
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
+	r = upfile()
+	PRINT_CALLBACK("ok,", r)
+	#remove tmp
+	run(f"rm -r {OUTDIR}", shell=True)
+	return True
+
+
 
 def getdur(url):
+	global ABORT
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
 	return int(float(check_output(f"ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 '{url}'", shell=True).decode("utf-8").strip()))
 
 def extract_middle_frame():
-    global URL
-    global OUTDIR
-    cap = VideoCapture(URL)
-    if not cap.isOpened():
-        raise ValueError(f"Error: Cannot open video file {URL}.")
-    frame_count = int(cap.get(CAP_PROP_FRAME_COUNT))
-    middle_frame = frame_count // 2
-    cap.set(CAP_PROP_POS_FRAMES, middle_frame)
-    ret, frame = cap.read()
-    if ret:
-        output_path = f"{OUTDIR}/middle_frame.webp"
-        imwrite(output_path, frame, [int(IMWRITE_WEBP_QUALITY), CV2_WEBP])
-        print(f"Middle frame saved to {output_path}")
-    else:
-        raise ValueError("Error: Could not read the middle frame.")
-    cap.release()
-
+	global ABORT
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
+	global URL
+	global OUTDIR
+	global PRINT_CALLBACK
+	cap = VideoCapture(URL)
+	if not cap.isOpened():
+		PRINT_CALLBACK(f"Error: Cannot open video file {URL}.")
+		return False
+	frame_count = int(cap.get(CAP_PROP_FRAME_COUNT))
+	middle_frame = frame_count // 2
+	cap.set(CAP_PROP_POS_FRAMES, middle_frame)
+	ret, frame = cap.read()
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
+	if ret:
+		output_path = f"{OUTDIR}/middle_frame.webp"
+		imwrite(output_path, frame, [int(IMWRITE_WEBP_QUALITY), CV2_WEBP])
+		PRINT_CALLBACK(f"Middle frame saved to {output_path}")
+	else:
+		PRINT_CALLBACK("Error: Could not read the middle frame.")
+		return False
+	cap.release()
+	return True
 def generate():
+	global ABORT
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
 	global OUTDIR
 	global URL
 	global DUR
+	global PRINT_CALLBACK
 	# segment and durations
 	segment_duration = 3 if DUR > 15 else DUR / 5
 	start = int(DUR*0.1)
@@ -60,6 +122,7 @@ def generate():
 		(f"{FFMPEG} -ss {end} -i '{URL}' -t {segment_duration} -vf scale=320:320 {EXTRA_MP4} -filter:v fps=fps=6 -an '{OUTDIR}/end.mp4'", "end")
 	]
 	for cmd, name in commands:
+		if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
 		thread = Thread(target=lambda c: run(c, shell=True), args=(cmd,), name=name)
 		threads.append(thread)
 		thread.start()
@@ -71,32 +134,19 @@ def generate():
 	partcount = allmp4.count(".mp4'")
 	filter_complex = " ".join([f"[{i}:v]" for i in range(partcount)])
 	cmd = f"{FFMPEG} {allmp4} -filter_complex '{filter_complex} concat=n={partcount}:v=1 [v]' -map '[v]' -vcodec libwebp {EXTRA_WEBP} -lossless 0 '{OUTDIR}/out.webp'"
-	print("running cmd:", cmd)
+	PRINT_CALLBACK("running cmd:", cmd)
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
 	run(cmd, shell=True)
 	# remove temp
 	run(f"rm '{OUTDIR}/start.mp4' '{OUTDIR}/start_1.mp4' '{OUTDIR}/middle.mp4' '{OUTDIR}/middle_1.mp4' '{OUTDIR}/end.mp4'", shell=True)
-	print("done")
+	PRINT_CALLBACK("done")
+	return  True
 
 def upfile():
-	# server-side thingies
-	return ""
-if __name__ == "__main__":
-	URL = fr"{argv[1]}"
-	WEBURL = fr"{argv[2]}"
-	FTYPE = fr"{argv[3]}"
-	DUR = getdur(URL)
-	randomname = str(abs(hash(URL)))
-	with open(r"./config/config.json", "r") as f:
-		j = load(f)
-		OUTDIR = fr'{j["tmp_dir"]}/{randomname}'
-		APIKEY = j["api_key"]
-	if not os_path.exists(OUTDIR):
-		makedirs(OUTDIR)
-	print(f"url: {URL} && got dur: {DUR} && && outdir: {OUTDIR}")
-	generate()
-	extract_middle_frame()
-	r = upfile()
-	print("ok,", r)
-	#remove tmp
-	run(f"rm -r {OUTDIR}", shell=True)
-print("\n\r")
+	global ABORT
+	if ABORT: print("!!!abort signal!!!");return 0, "abort signal"
+	#TODO: server-side thingies
+
+	#for now, lets use outside script
+	r = run(f"", shell=True)
+	return r
